Moved schema validation in its own file
